Добавлен подсчет стоимости с учетом скидки

+++ src/main/java/com/school/things/dto/PriceServiceListDTO.java
@@ -11,11 +11,11 @@ import lombok.ToString;
 @Builder
 public class PriceServiceListDTO {
     private Long id;
-    private int cost;
+    private double cost;
     private Long schoolServiceId;
     private String schoolServiceName;
 
-    public PriceServiceListDTO(Long id, int cost, Long schoolServiceId, String schoolServiceName) {
+    public PriceServiceListDTO(Long id, double cost, Long schoolServiceId, String schoolServiceName) {
         this.id = id;
         this.cost = cost;
         this.schoolServiceId = schoolServiceId;
